Refactor danger variables of dangerfinder
