+++ roboteam_ai/src/DangerFinder/modules/CanShootModule.h
@@ -13,10 +13,8 @@ namespace dangerfinder {
  */
 
 class CanShootModule : public DangerModule {
- private:
-  double canShootDangerScore;
 public:
-	CanShootModule(double canShootDangerScore);
+	explicit CanShootModule(double danger);
 	PartialResult calculate(const roboteam_msgs::WorldRobot& bot, const roboteam_msgs::World& world) override;
 };
