+++ roboteam_ai/src/DangerFinder/modules/DangerModule.cpp
@@ -4,18 +4,7 @@ namespace rtt {
 namespace ai {
 namespace dangerfinder {
 
-PartialResult::PartialResult() : score(0), flags(0) {}
-PartialResult::PartialResult(double score, DangerFlag flags) : score(score), flags(flags) {}
-
-PartialResult& PartialResult::operator+=(const PartialResult& b) {
-	score += b.score;
-	flags |= b.flags;
-	return *this;
-}
-
-PartialResult operator+(PartialResult a, PartialResult b) {
-	return { a.score + b.score, (DangerFlag)(a.flags | b.flags) };
-}
+DangerModule::DangerModule(double danger) : danger(danger) { }
 
 } // dangerfinder
 } // ai
