+++ roboteam_ai/src/DangerFinder/modules/FreeModule.cpp
@@ -6,7 +6,7 @@ namespace rtt {
 namespace ai {
 namespace dangerfinder {
 
-FreeModule::FreeModule(double danger) : danger(danger) { }
+FreeModule::FreeModule(double danger) : DangerModule(danger) { }
 
 PartialResult FreeModule::calculate(const roboteam_msgs::WorldRobot& bot, const roboteam_msgs::World& world) {
 	Vector2 ballPos(world.ball.pos);
