--- roboteam_ai/src/DangerFinder/modules/WillReceiveBallModule.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "DangerModule.h"
-
-namespace rtt {
-namespace ai {
-namespace dangerfinder {
-
-class WillReceiveBallModule : public DangerModule {
-public:
-    WillReceiveBallModule() = default;
-    PartialResult calculate(const roboteam_msgs::WorldRobot& bot, const roboteam_msgs::World& world) override;
-};
-
-} // dangerfinder
-} // ai
-} // rtt
